test(artifacts): Improve artifact tests (#672)
* test(artifacts): Update artifact tests

Given that I've just updated a bunch of places where we are using
artifacts, now is a good time to add some tests on some of the
behavior consumers are expecting but that there are no tests for.

In particular:
* Improve the serialization/deserialization tests by separately
testing these two and by using JsonNodeFactory to create JSON
to test against.
* Add a lot more tests on the behavior of metadata as it's a bit
complex and worth explicitly checking that lombok/Jackson are
doing what we expect there. (Particularly to have test in palce
to make sure nothing breaks when we later remove the deprecated
methods.)
To ensure that toBuilder properly handles metadata, add a test.

* test(artifacts): Improve ExpectedArtifact tests

Separately test serialization/deserialization and use JsonNodeFactory
to generate actual JSON strings.

* test(artifacts): Convert groovy tests to Java

The only groovy in this entire module is a few tests; convert them
to Java.

* refactor(artifacts): Clean up build.grade

Now that there is no groovy left (including in tests) we can remove
groovy-related dependencies. Also clean up the junit config to use
the junit-jupiter-engine.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>