fix(artifacts): Fix NPE deserializing artifacts (#676)
It turns out that the Nonnull annotation I added to metadata caused
lombok to insert an overzealous null-check in the .metdata() function
in the artifact builder, which breaks deserializing artifacts with
the metadata explicitly set to null.

Let's explicitly write out the Lombok-generated function, and instead
of throwing an NPE, just translate the null value to an empty Hashmap.

The added test failed before this change and passes after.