fix(perf): retrievePipelinesForPipelineConfigIdsBetweenBuildTimeBound… (#3738)
* fix(perf): retrievePipelinesForPipelineConfigIdsBetweenBuildTimeBoundary shouldn't get all executions

Because of "elaborate" ability to filter in `searchForPipelinesByTrigger` it must combine the filtering capabilities of execution repository and plain code.
To do that, it would retrieve ALL executions first and then filter to the subset desired (default 10).

This change adds filter on a page-by-page execution retrieval, so only the minimum number of results are retrieved from the DB

Ideally, this endpoint just goes away. But this is the minimum change to make it palatable perf wise

* fixup! fix(perf): retrievePipelinesForPipelineConfigIdsBetweenBuildTimeBoundary shouldn't get all executions

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>