fix(kubernetes): Fail with a useful message if manifest not found (#4690)
* refactor(kubernetes): Clean up EnableDisableManifestDescription

Mostly the goal here is to make loadBalancers Nonnull, which
required overriding the lombok-generated accessor so that we can
translate a null that's passed in to an empty list. While here,
make the list immutable.

Also add some tests, particularly on deserialization.

I did *not* make the class immutable because one must draw some
line at which to stop refactoring.

* refactor(kubernetes): Make KubernetesManifestTraffic null-safe

Also make it immutable, and add some tests.

* fix(kubernetes): Fail with a useful message if manifest not found

This commit adds nullity annotations to a number of functions
related to the disable manifest stage so that we can better
understand what can be null.

It then throws a useful error message when a requested manifest
is null, preventing what was a downstream error far from the
actual issue (and that had less information about the error).

* fix(kubernetes): Add missing parenthesis

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>