fix(log noise - mptv2): denoise some MPTv2 errors (#3770)
Currently, if there are any templated pipelines that are not fully specified (which is very common as users are iterating on pipeline design) planning those pipelines will generate lots of noise logs.
This is expecially annoying when this is done as part of dependentpipeline starter and so it looks like an exception occured during an execution when in fact it had nothing to do with this particular execution.

This change silences those logs for dependent pipeline starter, there will be a part 2 of this change where better logic is applied to `/v2/pipelineTemplates/plan` to not log exceptions for similar cases (this EP is called by echo alot)

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>